Refactor PR so that DSL is completely separate from future pystorm
